Refactor Action
