--- .editorconfig
@@ -1,11 +1,0 @@
-root = true
-
-[*]
-charset = utf-8
-end_of_line = lf
-insert_final_newline = true
-trim_trailing_whitespace = true
-
-[*.{js,json,ts,yml}]
-indent_style = space
-indent_size = 2
