+++ .github/workflows/ci.yml
@@ -1,0 +1,58 @@
+name: CI
+
+on:
+  push:
+    branches:
+      - main
+  pull_request:
+    branches:
+      - main
+
+env:
+  registry: ghcr.io
+  repository: '${{ github.repository }}'
+  tag: 'test-${{ github.run_id }}'
+  image: 'ghcr.io/${{ github.repository }}:test-${{ github.run_id }}'
+  username: '${{ github.repository_owner }}'
+  password: '${{ github.token }}'
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - run: yarn install
+
+      - name: Execute unit tests
+        run: yarn test
+
+      - name: Log In to GitHub Container Registry
+        uses: docker/login-action@v1
+        with:
+          registry: '${{ env.registry }}'
+          username: '${{ env.username }}'
+          password: '${{ env.password }}'
+
+      - name: Add example image to container registry for test repository
+        run: |
+          docker pull hello-world
+          docker tag hello-world ${{ env.image }}
+          docker push ${{ env.image }}
+
+      - name: Add One Tag
+        uses: ./
+        with:
+          registry: '${{ env.registry }}'
+          repository: '${{ env.repository }}'
+          target: '${{ env.tag }}'
+          tags: '${{ env.tag }}-single'
+
+      - name: Add Multiple Tags
+        uses: ./
+        with:
+          registry: '${{ env.registry }}'
+          repository: '${{ env.repository }}'
+          target: '${{ env.tag }}'
+          tags: |
+            ${{ env.tag }}-multiple-a
+            ${{ env.tag }}-multiple-b
