--- .github/workflows/package-action.yml
@@ -1,24 +1,0 @@
-name: "Package Action"
-
-on:
-  push:
-    branches:
-      - "release"
-
-jobs:
-  push-changes:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-        with:
-          ssh-key: "${{ secrets.COMMIT_KEY }}"
-      - name: Install Javascript dependencies with npm
-        run: npm install
-      - name: Package action for distribution
-        run: npm run package
-      - name: Push packaged action to branch
-        uses: EndBug/add-and-commit@v7
-        with:
-          add: "dist --force"
-          default_author: github_actions
-          message: "build: Package action as `dist` with latest changes"
