--- .github/workflows/sync-release-branch.yml
@@ -1,23 +1,0 @@
-name: "Synchronise To Release Branch"
-
-on:
-  push:
-    branches:
-      - "main"
-
-env:
-  source: "origin/main"
-  release: "release"
-
-jobs:
-  update-branch:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-        with:
-          ref: "${{ env.release }}"
-          fetch-depth: 0
-          ssh-key: "${{ secrets.COMMIT_KEY }}"
-      - uses: pr-mpt/actions-merge-branch@v2
-        with:
-          from: "${{ env.source }}"
