--- .github/workflows/tests-unit.yml
@@ -1,15 +1,0 @@
-name: "Test Action (Unit)"
-
-on:
-  push:
-    branches:
-      - "**"
-
-jobs:
-  test-unit:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-      - run: npm install
-      - name: Execute unit tests
-        run: npm run test
