--- .github/workflows/validate-tag.yml
@@ -1,22 +1,0 @@
-name: "Limit Tags To Distributable Commits"
-
-on:
-  push:
-    tags:
-      - "**"
-
-jobs:
-  validate-tag:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-        with:
-          ssh-key: "${{ secrets.COMMIT_KEY }}"
-      - name: Test that the tagged commit includes `dist`
-        uses: pr-mpt/actions-assert@v2
-        with:
-          assertion: npm://@assertions/directory-exists
-          expected: dist
-      - if: failure()
-        name: Delete invalid tag
-        uses: pr-mpt/actions-delete-tag@v1
