+++ .gitignore
@@ -1,4 +1,4 @@
-/lib
-/node_modules
-npm-debug.log*
 /dist
+/node_modules
+/coverage
+yarn.lock
