+++ package.json
@@ -1,40 +1,36 @@
 {
-  "name": "@shrinks/actions-docker-registry-tag",
-  "description": "Add many tags to an image in a Docker Registry",
+  "name": "update-registry-tag",
+  "description": "Update Docker Registry",
   "main": "lib/main.js",
   "scripts": {
-    "compile": "tsc",
-    "format": "prettier --write **/*.ts",
+    "build": "tsc",
     "lint": "eslint src/**/*.ts",
     "package": "npm run compile && ncc build --source-map --license licenses.txt",
-    "test": "jest",
-    "style": "npm run format && npm run lint"
+    "test": "jest --coverage"
   },
   "keywords": [
     "actions",
     "github-actions",
     "docker"
   ],
-  "author": "Samuel Ryan <sam@samryan.co.uk>",
-  "license": "MIT",
+  "author": "Theia Scientific, LLC",
   "dependencies": {
     "@actions/core": "^1.2.6",
     "@types/node-fetch": "^2.5.10",
-    "node-fetch": "^2.6.1"
+    "node-fetch": "^3.2.10"
   },
   "devDependencies": {
-    "@types/jest": "^26.0.15",
-    "@types/node": "^14.14.9",
-    "@typescript-eslint/parser": "^4.8.1",
-    "@vercel/ncc": "^0.25.1",
-    "eslint": "^7.17.0",
+    "@types/jest": "^28.1.8",
+    "@types/node": "^18.7.13",
+    "@typescript-eslint/parser": "^5.35.1",
+    "@vercel/ncc": "^0.34.0",
+    "eslint": "^8.22.0",
     "eslint-plugin-github": "^4.1.1",
-    "eslint-plugin-jest": "^24.1.3",
-    "jest": "^26.6.3",
-    "jest-circus": "^26.6.3",
-    "js-yaml": "^3.14.0",
-    "prettier": "2.2.1",
-    "ts-jest": "^26.4.4",
+    "eslint-plugin-jest": "^26.8.7",
+    "jest": "^29.0.1",
+    "jest-circus": "^29.0.1",
+    "js-yaml": "^4.1.0",
+    "ts-jest": "^28.0.8",
     "typescript": "^4.1.3"
   }
 }
